Finished integration, along with some bug fixes.
git-svn-id: svn://acme.able.cs.cmu.edu/rainbow/v2/trunk@1423 eea7622e-ad16-0410-96ee-d91cb3d3ee64

+++ rainbow-brass/src/main/java/org/sa/rainbow/brass/model/map/BatteryPredictor.java
@@ -4,18 +4,18 @@ package org.sa.rainbow.brass.model.map;
 
 public class BatteryPredictor {
 
-	
-	public float batteryConsumption (String speed, float time){
-		float base_consumption;
-		if (speed.equals("HALF_SPEED")){ 
-			base_consumption = 1.674f*time+287.5f;
-		}
-		else{ 
-			base_consumption = 3.89f*time+582.6f;
-		}
-		float kinect_consumption = 1.0386f * time;
-		float nuc_consumption = 3.08f * time;
-		return base_consumption + kinect_consumption + nuc_consumption;
-	}
-	
+
+    public double batteryConsumption (String speed, double time) {
+        double base_consumption;
+        if (speed.equals("HALF_SPEED")){ 
+            base_consumption = 1.674f*time+287.5f;
+        }
+        else{ 
+            base_consumption = 3.89f*time+582.6f;
+        }
+        double kinect_consumption = 1.0386f * time;
+        double nuc_consumption = 3.08f * time;
+        return base_consumption + kinect_consumption + nuc_consumption;
+    }
+
 }
